move mod and tests into one file
